Expand running status in situ
Alter the parser so that the result of parsing the last message is passed to the next.  Use this to determine the parentage of any running status message and fail if the parentage is unsuitable.  Translate the running status to the channel message and channel type of the parent.

Remove translateRunningStatus which used to do this as a post-processing step.